Removed failing test for double-slash in object key names
Removed failing test which expected double-slash characters to be retained
in S3 object key names, e.g. "this-is//a-key-name". The literal double-slash
is replaced with a single slash by the underlying HttpClient library and this
isn't worth fighting for, so key names with this edge case will not be
supported.